% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_com.R
\name{arrow_a}
\alias{arrow_a}
\title{Align a highlighted region's assignment operators.}
\usage{
arrow_a()
}
\value{
Aligns the single assignment operators (\code{<-}) within a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
