% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{check_before_run_query}
\alias{check_before_run_query}
\title{Verify input parameters before running the query to the TAXREF API}
\usage{
check_before_run_query(sp, lang = "sci", group_fr = NULL)
}
\arguments{
\item{sp}{A character string or list of character strings corresponding to the name(s) of a taxon (species or other).}

\item{lang}{A character string defining whether 'sp' is a vernacular ('fr') or scientific ('sci') name.
Default value is 'sci'.}

\item{group_fr}{A character string defining if a filter must be made on the taxon during the query (amphibians,
mammals, birds...) If specified, lang must be set to 'fr'.}
}
\value{
Returns TRUE if the input parameters are correctly formatted and compatible with each other, and stops
the function with an error message otherwise.
}
\description{
This function verifies that the input parameters sp, lang, and group_fr are correctly formatted and compatible
with each other before running the query to the TAXREF API.
}
\examples{
\dontrun{
check_before_run_query(sp = "Puma concolor", lang = "sci")
check_before_run_query(sp = c("Puma concolor", "Anas platyrhynchos"), lang = "sci")
check_before_run_query(sp = "Puma concolor", lang = "fr", group_fr = "Mammifères")
check_before_run_query(sp = "Puma concolor", lang = "fr")
check_before_run_query(sp = "Puma concolor", lang = "fr", group_fr = NULL)
check_before_run_query(sp = "Puma concolor", lang = "sci", group_fr = "Mammifères")
}

}
