% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{clean_loop_sampled}
\alias{clean_loop_sampled}
\title{Clean a loop}
\usage{
clean_loop_sampled(path_sf, clean)
}
\arguments{
\item{path_sf}{An sf object representing a loop.}

\item{clean}{Logical. If TRUE, the function will clean the loop. If FALSE, the function will return the loop without cleaning it.}
}
\value{
An sf object representing a cleaned loop.
}
\description{
This function cleans a loop by eliminating any round trips.
}
