% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_com.R
\name{com_a}
\alias{com_a}
\title{Align a highlighted region's comment operators.}
\usage{
com_a()
}
\value{
Aligns the comment sign (\code{#}) within a highlighted region.
}
\description{
Align a highlighted region's comment operators.
}
