% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_fr.R
\name{date_fr}
\alias{date_fr}
\title{Formats date in French format}
\usage{
date_fr(date, sep = "/")
}
\arguments{
\item{date}{a date object to format in French format}

\item{sep}{character separator. Default is "/".}
}
\value{
a character vector of the date reformatted in French format
}
\description{
This function takes a date as input and reformats it in French format (dd/mm/yyyy)
}
