% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{find_cd_nom}
\alias{find_cd_nom}
\title{Find the CD_NAME of a species from single or liste of taxon name (use the taxref API)}
\usage{
find_cd_nom(sp, lang = "sci")
}
\arguments{
\item{sp}{is a character string corresponding to the name of a taxon (species or other).}

\item{lang}{defines whether 'sp' is a vernacular (fr) or scientific (latin) name}
}
\description{
Find the CD_NAME of a species from single or liste of taxon name (use the taxref API)
}
