% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{find_shortest_loops}
\alias{find_shortest_loops}
\title{Find the shortest paths between the nodes of a graph, as a loop}
\usage{
find_shortest_loops(g, selected_nodes)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{selected_nodes}{a vector of nodes}
}
\value{
a data frame containing the shortests paths
}
\description{
This function takes a graph and a list of nodes as input, and returns a data frame containing the shortest paths
between each pair of nodes in the list.
}
