% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_validity.R
\name{fix_geom}
\alias{fix_geom}
\title{Fix geom with cleangeo buffer methode for sf object}
\usage{
fix_geom(poly)
}
\arguments{
\item{poly}{is a sf object (polygon or multipolygos).}
}
\description{
Fix geom with cleangeo buffer methode for sf object
}
