% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{g_filter}
\alias{g_filter}
\title{Filter a graph by distance from a starting node}
\usage{
g_filter(g, start_node, dist)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{start_node}{the starting node}

\item{dist}{the maximum distance to consider}
}
\value{
the filtered graph
}
\description{
This function takes a graph and a start node as input, and returns a subgraph consisting of the nodes at a certain
distance from the start node.
}
