% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_validity.R
\name{geom_is_valid}
\alias{geom_is_valid}
\title{Checks the validity of the geometry}
\usage{
geom_is_valid(layer)
}
\arguments{
\item{layer}{is a sf object.}
}
\description{
Checks the validity of the geometry
}
