% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{H1}
\alias{H1}
\title{Defines the style of a level 1 title - default version}
\usage{
H1(text, even = T, nchartext, max_length, char)
}
\arguments{
\item{text}{is a title}

\item{even}{is a boolean. the text has an even number of letters (TRUE). The text has an odd number of letters (FALSE)}

\item{nchartext}{is an integer vector. Is the number of characters in the title.}

\item{max_length}{is an integer vector. Corresponds to the size that will be generated.}

\item{char}{is fill character. "#" by default.}
}
\description{
Defines the style of a level 1 title - default version
}
