% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{list_group_fr}
\alias{list_group_fr}
\title{List available taxon groups for query to TAXREF API}
\usage{
list_group_fr()
}
\value{
A character vector containing the available taxon groups.
}
\description{
This function lists the available taxon groups that can be used to filter the query to the TAXREF API.
}
