% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{multiple_sample_in_net}
\alias{multiple_sample_in_net}
\title{Sample paths or loops from a graph}
\usage{
multiple_sample_in_net(
  g,
  n,
  min_length,
  max_length,
  clean,
  overlap_check,
  length_check,
  loop
)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{n}{the number of paths to sample}

\item{min_length}{the minimum length for a path}

\item{max_length}{the maximum length for a path}

\item{clean}{logical indicating whether to clean the paths or not}

\item{overlap_check}{logical indicating whether to check for overlap between paths}

\item{length_check}{logical indicating whether to check the length of the paths}

\item{loop}{logical indicating whether the generated path should be looped or on its way.}
}
\value{
a data frame containing the sampled paths
}
\description{
This function calls the sampling_loop function several times in a row to produce several random paths paths,
checking that these paths have no common parts and that their length is within a specified range.
}
