% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{nchar_is_even}
\alias{nchar_is_even}
\title{Boolean function to check if a number is even or odd}
\usage{
nchar_is_even(nchartext)
}
\arguments{
\item{nchartext}{is an integer vector.}
}
\description{
Boolean function to check if a number is even or odd
}
