% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{sample_in_network}
\alias{sample_in_network}
\title{Sample one or multiple paths or loops from a graph}
\usage{
sample_in_network(
  g,
  n = 1,
  min_length,
  max_length,
  overlap_check = FALSE,
  clean = TRUE,
  length_check = TRUE,
  loop = FALSE
)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{n}{the number of loops to sample, default is 1}

\item{min_length}{the minimum length for a way (in units of crs)}

\item{max_length}{the maximum length for a way (in units of crs)}

\item{overlap_check}{logical indicating whether to check for overlap between ways}

\item{clean}{logical indicating whether to clean the loops or not (only work for loop)}

\item{length_check}{logical indicating whether to check the length of the ways}

\item{loop}{logical indicating whether the generated way should be looped or from point A to point B.}
}
\value{
a data frame containing the sampled loops
}
\description{
This function samples one or more ways (loop or path) of a graph by selecting random nodes as starting points and
finding the shortest way between them. The length of the way is then checked to ensure that it is within the
specified range, and the resulting path can also be cleaned up to remove round trips. In
addition, the function can check for overlap between ways, if desired.
}
