% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{sample_one_node}
\alias{sample_one_node}
\title{Sample a starting node from a graph (for loops)}
\usage{
sample_one_node(g)
}
\arguments{
\item{g}{an object of class sfnetwork}
}
\value{
the randomly selected node
}
\description{
This function selects a node at random from a graph and returns it.
}
\examples{
\dontrun{
sample_one_node(g)
}
}
