% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{sample_two_nodes}
\alias{sample_two_nodes}
\title{Sample a start and end point on the network (for paths)}
\usage{
sample_two_nodes(g, nb_nodes = 2)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{nb_nodes}{nb_nodes is the number of total points (start and finish). The starting point is always equal to 1.
Only the arrival points can be more than 1}
}
\value{
a list with the starting point (from) and the ending point(s) (to)
}
\description{
This function selects two nodes at random from a graph and returns it.
}
