% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{sampling_path}
\alias{sampling_path}
\title{Sample and return a path from a graph}
\usage{
sampling_path(g)
}
\arguments{
\item{g}{an object of class sfnetwork}
}
\value{
an sf object representing the sampled path
}
\description{
Sample and return a path from a graph
}
