% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wax_setup.R
\name{setup_folders}
\alias{setup_folders}
\title{Create folders structure for new R project.}
\usage{
setup_folders(mode = "default")
}
\arguments{
\item{mode}{character indicating the type of structure to be created. Default is "default".}
}
\description{
This function creates a directory structure for new R project.
}
