% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_network.R
\name{shortest_loops_to_sf}
\alias{shortest_loops_to_sf}
\title{Transform shortest paths into sf format}
\usage{
shortest_loops_to_sf(g, shortest_paths_df)
}
\arguments{
\item{g}{an object of class sfnetwork}

\item{shortest_paths_df}{a data frame containing the shortest paths}
}
\value{
an sf object representing the shortest paths
}
\description{
This function takes a graph and a data frame of paths as input, and returns an object of class sf containing the
paths of the data frame.
}
