% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wax.R
\name{theme_wax}
\alias{theme_wax}
\title{Customize the theme of a ggplot2 plot using the wax_palette function}
\usage{
theme_wax(
  theme = "defaut",
  key = list(defaut = list(base_size = 13, family_title = "sans", family_text = "sans",
    primary_color = "#33302C", secondary_color = "#59544e", color_grid = "#59544e"), wax
    = list(base_size = 13, family_title = "roboto-slab", family_text = "fira-sans",
    primary_color = "#33302C", secondary_color = "#59544e", color_grid = "#E8E3DC"))
)
}
\arguments{
\item{theme}{the overall style of the chart. Can be either "defaut" or "wax".}

\item{key}{a list of predefined styles.}
}
\value{
a ggplot2 theme object.
}
\description{
This function allows the user to customize the overall style of a ggplot2 plot by using the `wax_palette` function.
The function takes a number of arguments to control the appearance of the plot, including the base font size,
the font families for the plot title and text, the primary and secondary colors, and the color of the grid lines.
The function also has a predefined set of styles that can be selected using the `key` argument.
}
