% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{use_taxref_api}
\alias{use_taxref_api}
\title{Send a query to the TAXREF API}
\usage{
use_taxref_api(url)
}
\arguments{
\item{url}{is there query formulate in url format}
}
\description{
This function sends a query in URL format to the TAXREF API and returns the response in JSON format.
It can be used to retrieve taxonomic information from the database.
}
