% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_wax.R
\name{wax_pal}
\alias{wax_pal}
\title{Prepare color palette like viridis methode}
\usage{
wax_pal(
  n,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "c_single_hue"
)
}
\arguments{
\item{n}{is the number of colors to be in the palette.}

\item{alpha}{is the alpha transparency, a number in [0,1], see argument alpha in hsv.}

\item{begin}{is the (corrected) hue in [0,1] at which the color map begins.}

\item{end}{is the (corrected) hue in [0,1] at which the color map ends.}

\item{direction}{sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{is choice of palette in wax_colours data}
}
\description{
Prepare color palette like viridis methode
}
