% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Clean a data frame}
\usage{
clean_df(df, remove_space = TRUE, lower_case = TRUE, latin_ascii = TRUE)
}
\arguments{
\item{df}{A data frame. The data frame to clean.}

\item{remove_space}{A logical indicating if spaces should be removed from the data frame. Default is TRUE.}

\item{lower_case}{A logical indicating if the data frame should be converted to lower case. Default is TRUE.}

\item{latin_ascii}{A logical indicating if the data frame should be converted to Latin-ASCII. Default is TRUE.}
}
\value{
a data frame with cleaned and formatted data.
}
\description{
This function takes a data frame as input and applies several cleaning operations on it. These operations include
removing spaces, converting to lower case, converting to Latin-ASCII, and cleaning column names using the function
`clean_colnames()`.
}
