% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Clean a data frame}
\usage{
clean_df(df, remove_space = TRUE, lower_case = TRUE, latin_ascii = TRUE)
}
\arguments{
\item{df}{is the data frame to clean}

\item{remove_space}{is a logical indicating if spaces should be removed from the data frame}

\item{lower_case}{is a logical indicating if the data frame should be converted to lower case}

\item{latin_ascii}{is a logical indicating if the data frame should be converted to Latin-ASCII}
}
\description{
Clean a data frame
}
