% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_df_by.R
\name{multiple_df_by}
\alias{multiple_df_by}
\title{Multiplies a dataframe by values}
\usage{
multiple_df_by(df, by)
}
\arguments{
\item{df}{is a dataframe.}

\item{by}{is liste of value (liste or column).}
}
\description{
Multiplies a dataframe by values
}
