% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{parse_url_for_taxref_api}
\alias{parse_url_for_taxref_api}
\title{Parse the url to make a query to the taxref api}
\usage{
parse_url_for_taxref_api(sp, lang)
}
\arguments{
\item{sp}{is a character string corresponding to the name of a taxon (species or other).}

\item{lang}{defines whether 'sp' is a vernacular (fr) or scientific (latin) name}
}
\description{
Parse the url to make a query to the taxref api
}
