% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cd_nom.R
\name{query_to_taxref_api}
\alias{query_to_taxref_api}
\alias{query_loop_to_taxref_api}
\title{send the query to the TAXREF API and formulate the response}
\usage{
query_to_taxref_api(sp, lang)

query_loop_to_taxref_api(sp, lang)
}
\arguments{
\item{sp}{is a character string corresponding to the name of a taxon (species or other).}

\item{lang}{defines whether 'sp' is a vernacular (fr) or scientific (latin) name}
}
\description{
send the query to the TAXREF API and formulate the response
}
