% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_wax.R
\name{scale_wax}
\alias{scale_wax}
\alias{scale_color_wax}
\alias{scale_fill_wax}
\alias{scale_fill_wax_d}
\alias{scale_fill_wax_c}
\alias{scale_color_wax_c}
\alias{scale_color_wax_d}
\title{Scale color from wax_palette}
\usage{
scale_wax(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "c_single_hue",
  aesthetics = "colour"
)

scale_color_wax(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "c_single_hue",
  aesthetics = "colour"
)

scale_fill_wax(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "c_single_hue",
  aesthetics = "fill"
)

scale_fill_wax_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = TRUE,
  option = "c_single_hue",
  aesthetics = "fill"
)

scale_fill_wax_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "c_single_hue",
  aesthetics = "fill"
)

scale_color_wax_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = FALSE,
  option = "c_single_hue",
  aesthetics = "colour"
)

scale_color_wax_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  discrete = TRUE,
  option = "c_single_hue",
  aesthetics = "colour"
)
}
\arguments{
\item{...}{is the information passed to the underlying functions. Auto-completed, please ignore.}

\item{alpha}{is the alpha transparency, a number in [0,1], see argument alpha in hsv.}

\item{begin}{is the (corrected) hue in [0,1] at which the color map begins.}

\item{end}{is the (corrected) hue in [0,1] at which the color map ends.}

\item{direction}{sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{discrete}{Operator defining the use of the palette (TRUE for discrete, FALSE for continuous).}

\item{option}{is choice of palette in wax_colours data}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with.}
}
\description{
Scale color from wax_palette
}
