% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wax.R
\name{theme_wax}
\alias{theme_wax}
\title{Wax theme for ggplot2
Personal theme for ggplot2}
\usage{
theme_wax(
  theme = "defaut",
  key = list(defaut = list(base_size = 13, family_title = "sans", family_text = "sans",
    primary_color = "#33302C", secondary_color = "#59544e", color_grid = "#59544e"), wax
    = list(base_size = 13, family_title = "roboto-slab", family_text = "fira-sans",
    primary_color = "#33302C", secondary_color = "#59544e", color_grid = "#E8E3DC"))
)
}
\arguments{
\item{theme}{sets the overall style of the chart}

\item{key}{is a list of predefined styles}
}
\description{
Wax theme for ggplot2
Personal theme for ggplot2
}
